+++ docs/source/change_log.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 
 v1.05 - 2025-10-28 - VMLC-PV
 -----------------------------
-- axBOtorchOptimizer: skipping center and sobol as "models" when creating the ax client instance, this is to make sure that the client has a trainable model.  
+- axBOtorchOptimizer: skipping center and sobol as "models" when creating the ax client instance, this is to make sure that the client has a trainable model. Fixed small error in best Sobol value print.  
 - SIMsalabimAgent: Remove the check that enforces left-to-right or right-to-left arrangement of the energy level offsets (differences in designated energy levels). This is not necessary, as a later check ensures that each parameter is specified only once.   
 - JVAgent: Added new exp_format ‘^QFLSL-?\d+$’ i.e. QFLSL plus an integer. This is to simulate QFLS vs Vext curves.  
-- RateEqModel: Removed a factor 2 in the p1s expression.  
+- RateEqModel: Removed a factor 2 in the p1s expression. Added N0 as a potential fit parameters.  
+- scipyOptimizer: Added optimize_least_squares method to perform optimization using the least squares method from scipy. This is useful for fitting problems and can provide better results than the standard scipy.optimize.minimize method for certain problems.
 
 v1.04 - 2025-09-25 - VMLC-PV
 -----------------------------